backend/campaign_helper/authentication/permissions.py
from rest_framework import permissions


class IsUserOrAdmin(permissions.BasePermission):
    def has_object_permission(self, request, view, obj):
        return request.user.is_staff or request.user == obj
